requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
